function that searches for a value in a sorted array of integers using the Jump search algorithm and time complexity of Jump search algo.

+++ 0x1E-search_algorithms/100-jump.c
@@ -15,7 +15,7 @@ int jump_search(int *array, size_t size, int value)
 
 	if (array)
 	{
-		printf("Value checked array[%d] = [%d]\n", (int)j_previous, array[j_previous]);
+		printf("Value checked array[%ld] = [%d]\n", j_previous, array[j_previous]);
 		while (1)
 		{
 			if (j_next > size - 1)
